dcp-815 move classes to ingest package
