+++ exporter/ingest/service.py
@@ -1,74 +1,21 @@
 import json
 import logging
-from dataclasses import dataclass
-from enum import Enum
-from typing import List, Dict, Callable
+from typing import Callable
 
 import polling
 import requests
 from hca_ingest.api.ingestapi import IngestApi
 
-
-@dataclass
-class TerraExportError:
-    message: str
-
-    def to_dict(self) -> Dict:
-        return {
-            "message": self.message,
-            "errorCode": -1,
-            "details": {}
-        }
+from exporter.ingest.export_job import ExportEntity, ExportJobState, ExportJob
 
 
-@dataclass
-class TerraExportEntity:
-    assay_process_id: str
-    errors: List[TerraExportError]
-
-    def to_dict(self) -> Dict:
-        """
-        converts to a JSON as represented in ingest-core API
-        """
-        return {
-            "status": ExportJobState.EXPORTED.value,
-            "context": {
-                "assayProcessId": self.assay_process_id
-            },
-            "errors": [e.to_dict() for e in self.errors]
-        }
-
-
-class ExportJobState(Enum):
-    EXPORTING = "EXPORTING"
-    EXPORTED = "EXPORTED"
-    DEPRECATED = "DEPRECATED"
-    FAILED = "FAILED"
-
-
-@dataclass
-class TerraExportJob:
-    job_id: str
-    num_expected_assays: int
-    export_state: ExportJobState
-    is_data_transfer_complete: bool
-
-    @staticmethod
-    def from_dict(data: Dict) -> 'TerraExportJob':
-        job_id = str(data["_links"]["self"]["href"]).split("/")[0]
-        num_expected_assays = int(data["context"]["totalAssayCount"])
-        is_data_transfer_complete = data["context"].get("isDataTransferComplete")
-        return TerraExportJob(job_id, num_expected_assays, ExportJobState(data["status"].upper()),
-                              is_data_transfer_complete)
-
-
-class TerraExportJobService:
+class IngestService:
     def __init__(self, ingest_client: IngestApi):
         self.ingest_client = ingest_client
         self.logger = logging.getLogger(__name__)
 
     def create_export_entity(self, job_id: str, assay_process_id: str):
-        assay_export_entity = TerraExportEntity(assay_process_id, [])
+        assay_export_entity = ExportEntity(assay_process_id, [])
         create_export_entity_url = self.get_export_entities_url(job_id)
         requests.post(create_export_entity_url, json.dumps(assay_export_entity.to_dict()),
                       headers={"Content-type": "application/json"}, json=True).raise_for_status()
@@ -92,9 +39,9 @@ class TerraExportJobService:
     def get_job_state(self, job_id: str) -> ExportJobState:
         return self.get_job(job_id).export_state
 
-    def get_job(self, job_id: str) -> TerraExportJob:
+    def get_job(self, job_id: str) -> ExportJob:
         job_url = self.get_job_url(job_id)
-        return TerraExportJob.from_dict(self.ingest_client.get(job_url).json())
+        return ExportJob.from_dict(self.ingest_client.get(job_url).json())
 
     def get_job_url(self, job_id: str) -> str:
         return self.ingest_client.get_full_url(f'/exportJobs/{job_id}')
