+++ exporter/terra/terra_exporter.py
@@ -4,10 +4,10 @@ from hca_ingest.api.ingestapi import IngestApi
 
 from exporter import utils
 from exporter.graph.graph_crawler import GraphCrawler
+from exporter.ingest.service import IngestService
 from exporter.metadata.resource import MetadataResource
 from exporter.metadata.service import MetadataService
 from exporter.terra.client import TerraClient
-from exporter.terra.terra_export_job import TerraExportJobService
 
 LOGGER_NAME = __name__
 
@@ -18,7 +18,7 @@ class TerraExporter:
                  metadata_service: MetadataService,
                  graph_crawler: GraphCrawler,
                  dcp_staging_client: TerraClient,
-                 job_service: TerraExportJobService):
+                 job_service: IngestService):
         self.ingest_client = ingest_client
         self.metadata_service = metadata_service
         self.graph_crawler = graph_crawler
